fix: update lecturer flow
refactored and fixed lecturer flow
